Finally tests are passing in ubuntu docker!
The problem was the pocl.icd that was created by conda.
The solution was to use the pocl.icd created by apt.

Note: this is a rebase of 85 commits!

+++ developer/docker/build_docker.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 
+if [[ ! $(basename "$(pwd)") = 'docker' ]]; then
+    echo 'This script should run in the developer/docker directory.'
+    exit 1
+fi
+
 cp ../../environment.yml .
-docker build -f Dockerfile .
+docker -v build -f Dockerfile -t registry.gitlab.com/kirianlab/reborn/ubuntu-20.04:latest .
 rm environment.yml
-# docker login registry.gitlab.com
-# docker tag <hash>
-# docker push
+
+if [[ "$1" == "push" ]]; then
+    docker login registry.gitlab.com
+    docker push registry.gitlab.com/kirianlab/reborn/ubuntu-20.04:latest
+fi
